Template update

+++ template.cpp
@@ -1,12 +1,12 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-#define rep(i, a, b) for(int i = a; i < (b); ++i)
+#define rep(i, a, b) for(ll i = a; i < (b); ++i)
 #define all(x) begin(x), end(x)
-#define sz(x) (int)(x).size()
+#define sz(x) (ll)(x).size()
 typedef long long ll;
-typedef pair<int, int> pii;
-typedef vector<int> vi;
+typedef pair<ll, ll> pii;
+typedef vector<ll> vi;
 
 int main() {
 	cin.tie(0)->sync_with_stdio(0);
